:fire: login page removed. now, login is available and working in the navbar login modal
