--- src/pages/HomePage/Login/index.js
@@ -1,58 +1,0 @@
-import { useState, useContext } from "react";
-import { AuthContext } from "../../../contexts/authContext";
-import { api } from "../../../api/api";
-import { useNavigate } from "react-router-dom";
-
-export function Login() {
-  const [form, setForm] = useState({
-    email: "",
-    password: "",
-  });
-
-  const navigate = useNavigate();
-
-  const { setLoggedInUser } = useContext(AuthContext);
-
-  function handleChange(e) {
-    setForm({ ...form, [e.target.name]: e.target.value });
-  }
-
-  async function handleSumit(e) {
-    e.preventDefault();
-
-    try {
-      const response = await api.post("/user/log-in", form);
-      setLoggedInUser({ ...response.data });
-      console.log(response.data);
-
-      localStorage.setItem("loggedInUser", JSON.stringify(response.data));
-      if (response.data.user.role === "ADMIN") {
-        navigate("/admin");
-      } else {
-        navigate("/");
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  return (
-    <form onSubmit={handleSumit}>
-      <label>Email:</label>
-      <input
-        type="email"
-        name="email"
-        value={form.email}
-        onChange={handleChange}
-      />
-      <label>Senha:</label>
-      <input
-        type="password"
-        name="password"
-        value={form.password}
-        onChange={handleChange}
-      />
-      <button type="submit">Entrar!</button>
-    </form>
-  );
-}
